pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loftai-patterns"
version = "0.1.0"
description = "Pattern mining & backtesting toolkit"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "matplotlib>=3.7",
  "scikit-learn>=1.3",
  "stumpy>=1.12",
  "umap-learn>=0.5; python_version >= '3.10'",
  "tqdm>=4.66"
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["patterns"]
